initial commit
this is a basic build hosting app for use with Drone CI. A key is given,
which you store in an environment variable for the repo. A script
grabbed and run in the build container will upload files to the running app, which hosts them on a specific page for the repo.