add tagging for app updates
tags are set by drone when a platform build (binary, docker, etc.) or
release is completed and available. It is named, and also includes info such as release date and commit. Clients ping a URL for the latest tag they
want, and once it turns true, use the info to pull the update from the
appropriate source. Set to be used in jfa-go for simple self-updating
binaries and release notifications for docker.